to do: remove duplicate genres, parse multiple genre strings, and finished movies with votes > 350,000

+++ JSONFinalReview/JSONFinalReview/Movie.cs
@@ -26,7 +26,7 @@ namespace JSONFinalReview
        public string content_rating     {get;set;}
        public int budget                {get;set;}
        public int title_year            { get; set; }
-       public int imdb_score { get; set; }
+       public double imdb_score { get; set; }
        public int movie_facebook_likes { get; set; }
     }
 }
